refactor(radio): rename utils to sorting to semantically scope sorting logic
