+++ assets/src/components/radio/rtts/index.ts
@@ -7,4 +7,4 @@ export * from "./useRttQueue"
 export * from "./useRttSelection"
 export * from "./useRttQueueTabs"
 export * from "./rttQueueReducer"
-export * from "./utils"
+export * from "./sorting"
